requirement.txt
django
djangorestframework
django-filter
uritemplate
